+++ src/gmprocess/waveform_processing/resample.py
@@ -20,7 +20,7 @@ def resample(st, new_sampling_rate=None, method=None, a=None, config=None):
             Configuration dictionary (or None). See get_config().
 
     Returns:
-        StationStream: With resampling applied.
+        StationStream: Stream with resampling applied.
     """
     if not st.passed:
         return st