+++ src/gmprocess/waveform_processing/taper.py
@@ -20,7 +20,7 @@ def taper(st, type="hann", width=0.05, side="both", config=None):
             Configuration dictionary (or None). See get_config().
 
     Returns:
-        StationStream: With the taper applied.
+        StationStream: Stream with the taper applied.
     """
     if not st.passed:
         return st